Uber Case Study
Uber claims their drivers who work as independent contractors receive the benefits of flexibility, independence and on-demand working hours. Despite of these benefits, Uber has been having on-going law suits surrounding the topic of unfair employee treatments. Some drivers of Uber claim that they are not being treated as employees of Uber and that they do not receive any employment benefits. This experiment aims to evaluate the effect of a mechanism test that looks at a policy change for Uber switching from independent contracted drivers to actual employees (drivers) of Uber.